% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyWeather.R
\name{runShinyWeather}
\alias{runShinyWeather}
\title{runShinyWeather}
\usage{
runShinyWeather()
}
\description{
This is the function that can be run by the user to open the ShinyWeather App
}
\examples{
runShinyWeather()

}
