% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWeather.R
\docType{package}
\name{ShinyWeather}
\alias{ShinyWeather}
\title{This package makes a Shiny App that shows the weather in the Netherlands in an
intuitive way for children}
\section{Shiny App}{

\emph{ShinyWeather} contains the code for the application
}

